+++ src/com/utbm/smallWorld/peuples/PeupleChercheur.java
@@ -19,7 +19,7 @@ public class PeupleChercheur extends Peuple {
 		nbUniteMax = 20;
 	}
 	
-	public int bonusDefense(Territoire t, Peuple attaquant) {
-		return t.has(Laboratoire.class) ? 2 : 0;
+	public double bonusDefense(Territoire t, Peuple attaquant) {
+		return t.has(Laboratoire.class) ? 2.0 : 0.0;
 	}
 }
